refactor: Modify the use of AcPower related UseCases
