+++ app/src/main/java/com/thoughtworks/carapp/domain/hvac/AcPowerStatusUseCase.kt
@@ -1,4 +1,4 @@
-package com.thoughtworks.carapp.domain.acmode
+package com.thoughtworks.carapp.domain.hvac
 
 import android.car.VehiclePropertyIds
 import android.car.hardware.property.CarPropertyManager
@@ -7,14 +7,22 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
-class GetAcAutoStatusUseCase @Inject constructor(
+class AcPowerStatusUseCase @Inject constructor(
     private val vehiclePropertyManager: VehiclePropertyManager
 ) {
-    operator fun invoke(): Flow<Boolean> {
+    fun getAcPowerStatusFlow(): Flow<Boolean> {
         return vehiclePropertyManager.getPropertyFlow(
-            VehiclePropertyIds.HVAC_AUTO_ON,
+            VehiclePropertyIds.HVAC_AC_ON,
             CarPropertyManager.SENSOR_RATE_ONCHANGE
         )
             .map { it?.value as Boolean }
+    }
+
+    fun setAcPowerStatus(value: Boolean) {
+        vehiclePropertyManager.setProperty(
+            VehiclePropertyIds.HVAC_AC_ON,
+            HvacAreas.ALL,
+            value
+        )
     }
 }
