+++ app/src/main/java/com/thoughtworks/carapp/domain/hvac/SetAcAutoStatusUseCase.kt
@@ -1,0 +1,17 @@
+package com.thoughtworks.carapp.domain.hvac
+
+import android.car.VehiclePropertyIds
+import com.thoughtworks.carapp.device.VehiclePropertyManager
+import javax.inject.Inject
+
+class SetAcAutoStatusUseCase @Inject constructor(
+    private val vehiclePropertyManager: VehiclePropertyManager
+) {
+    operator fun invoke(value: Boolean) {
+        vehiclePropertyManager.setProperty(
+            VehiclePropertyIds.HVAC_AUTO_ON,
+            HvacAreas.ALL,
+            value
+        )
+    }
+}
