+++ app/src/main/java/com/thoughtworks/carapp/presentation/carsetting/ac/AcViewModel.kt
@@ -1,9 +1,8 @@
 package com.thoughtworks.carapp.presentation.carsetting.ac
 
-import com.thoughtworks.carapp.domain.acmode.GetAcAutoStatusUseCase
-import com.thoughtworks.carapp.domain.acmode.GetAcPowerStatusUseCase
-import com.thoughtworks.carapp.domain.acmode.SetAcAutoStatusUseCase
-import com.thoughtworks.carapp.domain.acmode.SetAcPowerStatusUseCase
+import com.thoughtworks.carapp.domain.hvac.AcPowerStatusUseCase
+import com.thoughtworks.carapp.domain.hvac.GetAcAutoStatusUseCase
+import com.thoughtworks.carapp.domain.hvac.SetAcAutoStatusUseCase
 import com.thoughtworks.carapp.presentation.base.BaseViewModel
 import com.thoughtworks.carapp.presentation.base.Event
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -17,18 +16,17 @@ sealed interface SettingScreenEvent : Event {
 
 @HiltViewModel
 class AcViewModel @Inject constructor(
-    getAcPowerStatus: GetAcPowerStatusUseCase,
+    private val acPowerStatus: AcPowerStatusUseCase,
     getAcAutoStatus: GetAcAutoStatusUseCase,
-    private val setAcPowerStatus: SetAcPowerStatusUseCase,
     private val setAcAutoStatus: SetAcAutoStatusUseCase,
 ) : BaseViewModel<SettingScreenEvent>() {
-    val isAcPowerOn: StateFlow<Boolean> = getAcPowerStatus().stateWith(false)
+    val isAcPowerOn: StateFlow<Boolean> = acPowerStatus.getAcPowerStatusFlow().stateWith(false)
     val isAcAutoOn: StateFlow<Boolean> = getAcAutoStatus().stateWith(false)
 
     override fun handleEvents(event: SettingScreenEvent) {
         when (event) {
             SettingScreenEvent.SwitchAcPowerEvent -> {
-                setAcPowerStatus(isAcPowerOn.value.not())
+                acPowerStatus.setAcPowerStatus(isAcPowerOn.value.not())
             }
             SettingScreenEvent.SwitchAcAutoEvent -> {
                 setAcAutoStatus(isAcAutoOn.value.not())
